Removed KernelUtils as recommended
